Changes the customer creation logic
