+++ src/model/user.ts
@@ -1,25 +1,5 @@
-import Database from "../utils/database";
+import UserDb from '../utils/user_db';
 const _ = require('lodash');
-
-class UserDb {
-    static instance = null;
-
-    constructor() {
-        if (!UserDb.instance) {
-
-            const dbDef: Database = new Database(
-                process.env.TEST_DB || 'users.json',
-                { users: [] }
-            );
-
-            UserDb.instance = dbDef.Instance;
-        }
-    }
-
-    get Instance() {
-        return UserDb.instance;
-    }
-}
 
 /**
  * @swagger
@@ -29,6 +9,19 @@ class UserDb {
  *       name:
  *         type: string
  *       username:
+ *         type: string
+ */
+
+/**
+ * @swagger
+ * definition:
+ *   customer_create_tupple:
+ *     properties:
+ *       username:
+ *         type: string
+ *       name:
+ *         type: string
+ *       password:
  *         type: string
  */
 export default class ModelUser {
